Move types to separate folder
