+++ src/blocks/addMemeToTags.ts
@@ -1,4 +1,4 @@
-import { LocaleAwarePullRequest, PipelineConfig } from '../types';
+import { LocaleAwarePullRequest, PipelineConfig } from '../types/types';
 const fs = require('fs');
 import path from 'path';
 export async function addMemeToTags(pullRequest: LocaleAwarePullRequest, config: PipelineConfig) {
