+++ src/blocks/push-changes.ts
@@ -1,4 +1,4 @@
-import { LocaleAwarePullRequest, PipelineConfig } from '../types';
+import { LocaleAwarePullRequest, PipelineConfig } from '../types/types';
 import sgit from 'simple-git/promise';
 export async function pushChanges(): Promise<void> {
   if (process.env.TRAVIS_BRANCH) {
