+++ src/blocks/remove-pullrequests.ts
@@ -1,4 +1,4 @@
-import { LocaleAwarePullRequest, PipelineConfig } from '../types';
+import { LocaleAwarePullRequest, PipelineConfig } from '../types/types';
 import rimraf from 'rimraf';
 
 export async function removePullrequests(pullRequest: LocaleAwarePullRequest, config: PipelineConfig) {
